# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Gin 的快速启动项目",
        "title": "Gin Admin",
        "contact": {},
        "version": "v1.0.0"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "AuthAPI"
                ],
                "summary": "Login system with username and password",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-dto_LoginToken"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "AuthAPI"
                ],
                "summary": "Logout system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/menus": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "AuthAPI"
                ],
                "summary": "Query current user menus based on the current user role",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-po_Menus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/password": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "AuthAPI"
                ],
                "summary": "Change current user password",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthUpdatePasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh-token": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "AuthAPI"
                ],
                "summary": "Refresh current access token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-dto_LoginToken"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "AuthAPI"
                ],
                "summary": "Get current user info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-po_User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "AuthAPI"
                ],
                "summary": "Update current user info",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthUpdateUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
...
        "/api/v1/captcha/image": {
            "get": {
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "CaptchaAPI"
                ],
                "summary": "Response captcha image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Captcha ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Reload captcha image (reload=1)",
                        "name": "reload",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Captcha image"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
...
        "/api/v1/menus": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "MenuAPI"
                ],
                "summary": "Query menu tree data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "-",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pagination size, less than 0 is considered as unlimited quantity. default(20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Display name of menu",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pagination index. default(1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status of menu (disabled, enabled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "total number of items",
                        "name": "total",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Type of menu (catalog, menu, button)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include resources",
                        "name": "withResources",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResultList-po_Menu"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "MenuAPI"
                ],
                "summary": "Create menu record",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.MenuCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-po_Menu"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Result-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AuthUpdatePasswordReq": {
            "type": "object",
            "required": [
                "newPassword",
                "oldPassword"
            ],
            "properties": {
                "newPassword": {
                    "description": "New password (md5 hash)",
                    "type": "string"
                },
                "oldPassword": {
                    "description": "Old password (md5 hash)",
                    "type": "string"
                }
            }
        },
        "dto.AuthUpdateUserReq": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email of user",
                    "type": "string",
                    "maxLength": 128
                },
                "nickName": {
                    "description": "Name of user",
                    "type": "string",
                    "maxLength": 64
                },
                "phone": {
                    "description": "Phone number of user",
                    "type": "string",
                    "maxLength": 32
                },
                "wechat": {
                    "description": "Wechat account",
                    "type": "string",
                    "maxLength": 64
                }
            }
        },
...
        "dto.List-po_Menu": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.Menu"
                    }
                },
                "limit": {
                    "description": "pagination size, less than 0 is considered as unlimited quantity. default(20)",
                    "type": "integer"
                },
                "page": {
                    "description": "pagination index. default(1)",
                    "type": "integer"
                },
                "total": {
                    "description": "total number of items",
                    "type": "integer"
                }
            }
        },
...
        "dto.Login": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "Login password (md5 hash)",
                    "type": "string"
                },
                "username": {
                    "description": "Login name",
                    "type": "string"
                }
            }
        },
        "dto.LoginToken": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "description": "Access token (JWT)",
                    "type": "string"
                },
                "expires": {
                    "description": "Expired time (second)s",
                    "type": "integer"
                },
                "refreshToken": {
                    "description": "Refresh token (JWT)",
                    "type": "string"
                },
                "tokenType": {
                    "description": "Token type (Usage: Authorization=${token_type} ${access_token})",
                    "type": "string"
                }
            }
        },
        "dto.MenuCreateReq": {
            "type": "object",
            "required": [
                "status",
                "type"
            ],
            "properties": {
                "component": {
                    "description": "Component path of view",
                    "type": "string"
                },
                "extra": {
                    "description": "Extras",
                    "type": "object",
                    "additionalProperties": {}
                },
                "method": {
                    "description": "Http method of resource",
                    "type": "string"
                },
                "name": {
                    "description": "Display name of menu",
                    "type": "string",
                    "maxLength": 128
                },
                "parentId": {
                    "description": "Parent ID (From Menu.ID)",
                    "type": "string"
                },
                "path": {
                    "description": "Access path of menu",
                    "type": "string"
                },
                "rank": {
                    "description": "Rank for sorting (Order by desc)",
                    "type": "integer"
                },
                "redirect": {
                    "description": "Redirect path of menu",
                    "type": "string"
                },
                "status": {
                    "description": "Status of menu (enabled, disabled)",
                    "type": "string",
                    "enum": [
                        "disabled",
                        "enabled"
                    ]
                },
                "title": {
                    "description": "Menu title",
                    "type": "string"
                },
                "type": {
                    "description": "Type of menu (catalog menu, button)",
                    "type": "string",
                    "enum": [
                        "catalog",
                        "menu",
                        "button"
                    ]
                }
            }
        },
...
        "dto.Result-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.Result-dto_LoginToken": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/dto.LoginToken"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.Result-po_Menu": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/po.Menu"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.Result-po_Menus": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.Menu"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.Result-po_User": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/po.User"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.ResultList-po_Menu": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/dto.List-po_Menu"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "po.Menu": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "Child menus",
                    "allOf": [
                        {
                            "$ref": "#/definitions/po.Menus"
                        }
                    ]
                },
                "component": {
                    "description": "Component path of view",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Create time",
                    "type": "string"
                },
                "extra": {
                    "description": "Extra data for frontend",
                    "type": "object",
                    "additionalProperties": {}
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "method": {
                    "description": "Http method of resource",
                    "type": "string"
                },
                "name": {
                    "description": "Display name of menu",
                    "type": "string"
                },
                "parentId": {
                    "description": "Parent ID (From Menu.ID)",
                    "type": "string"
                },
                "path": {
                    "description": "Access path of menu",
                    "type": "string"
                },
                "rank": {
                    "description": "Rank for sorting (Order by desc)",
                    "type": "integer"
                },
                "redirect": {
                    "description": "Redirect path of menu",
                    "type": "string"
                },
                "roles": {
                    "$ref": "#/definitions/po.Roles"
                },
                "status": {
                    "description": "Status of menu (enabled, disabled)",
                    "type": "string"
                },
                "title": {
                    "description": "Menu title",
                    "type": "string"
                },
                "type": {
                    "description": "Type of menu (catalog, menu, button)",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Update time",
                    "type": "string"
                }
            }
        },
        "po.Menus": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/po.Menu"
            }
        },
        "po.Role": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code of role (unique)",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Create time",
                    "type": "string"
                },
                "description": {
                    "description": "Details about role",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.Menu"
                    }
                },
                "name": {
                    "description": "Display name of role",
                    "type": "string"
                },
                "rank": {
                    "description": "Rank for sorting",
                    "type": "integer"
                },
                "status": {
                    "description": "Status of role (disabled, enabled)",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Update time",
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.User"
                    }
                }
            }
        },
        "po.Roles": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/po.Role"
            }
        },
        "po.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "Avatar URL",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Create time",
                    "type": "string"
                },
                "description": {
                    "description": "Details about user",
                    "type": "string"
                },
                "email": {
                    "description": "Email of user",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "nickName": {
                    "description": "Name of user",
                    "type": "string"
                },
                "phone": {
                    "description": "Phone number of user",
                    "type": "string"
                },
                "realName": {
                    "description": "Real name of user",
                    "type": "string"
                },
                "roles": {
                    "description": "Roles of user",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.Role"
                    }
                },
                "status": {
                    "description": "Status of user (activated, freezed)",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Update time",
                    "type": "string"
                },
                "username": {
                    "description": "Username for login",
                    "type": "string"
                },
                "wechat": {
                    "description": "Wechat account",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}